added spring auto rest docs config
