+++ src/main/java/com/idforideas/pizzeria/security/CustomAuthorizationFilter.java
@@ -41,7 +41,7 @@ public class CustomAuthorizationFilter extends OncePerRequestFilter {
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
             throws ServletException, IOException {
-        if(request.getServletPath().equals("/api/v1/login") || request.getServletPath().equals("/api/v1/users/refresh-token")) {
+        if(request.getServletPath().equals("/api/v1/oauth/token") || request.getServletPath().equals("/api/v1/oauth/token/refresh")) {
             filterChain.doFilter(request, response);
         } else {
             String authorizationHeader = request.getHeader(AUTHORIZATION);
