+++ src/main/java/com/idforideas/pizzeria/security/CustomEnvironmentVariables.java
@@ -3,5 +3,7 @@ package com.idforideas.pizzeria.security;
 public interface CustomEnvironmentVariables {
 
     public static final String SECRET = "SECRET";
+    public static final String USER_TEST = "USER_TEST";
+    public static final String PWD_TEST = "PWD_TEST";
     
 }
